# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tickets/{id}/transfer": {
      "post": {
        "operationId": "Transfers_offer",
        "summary": "",
        "description": "Offer one of your tickets to someone by email.\n\n**Roles:** CUSTOMER",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferTransferDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Transfers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/transfers/mine": {
      "get": {
        "operationId": "Transfers_mine",
        "summary": "",
        "description": "Transfers you've sent and received.\n\n**Roles:** CUSTOMER",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Transfers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/transfers/{id}/cancel": {
      "post": {
        "operationId": "Transfers_cancel",
        "summary": "",
        "description": "Take back an offer that hasn't been accepted yet.\n\n**Roles:** CUSTOMER",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Transfers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/transfers/{id}/resend": {
      "post": {
        "operationId": "Transfers_resend",
        "summary": "",
        "description": "Email a fresh link (the previous one stops working).\n\n**Roles:** CUSTOMER",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Transfers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/transfers/preview": {
      "post": {
        "operationId": "Transfers_preview",
        "summary": "",
        "description": "What's being offered (no sign-in needed; the token is the proof). POST so the token stays out of URLs and logs.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Transfers"
        ]
      }
    },
    "/api/v1/transfers/accept": {
      "post": {
        "operationId": "Transfers_accept",
        "summary": "",
        "description": "Accept: signed in with the address the ticket was sent to. The ticket gets a new QR code.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          }
        },
        "tags": [
          "Transfers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/transfers/decline": {
      "post": {
        "operationId": "Transfers_decline",
        "summary": "",
        "description": "Decline (no sign-in needed).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Transfers"
        ]
      }
    },
...
    "/api/v1/venues": {
      "get": {
        "operationId": "Venues_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Venues"
        ]
      },
      "post": {
        "operationId": "Venues_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVenueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Venues"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ADMIN"
      }
    },
    "/api/v1/venues/{id}": {
      "get": {
        "operationId": "Venues_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Venues"
        ]
      }
    },
    "/api/v1/venues/{id}/sections": {
      "post": {
        "operationId": "Venues_createSection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Venues"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ADMIN"
      }
    },
    "/api/v1/venues/{id}/access-zones": {
      "post": {
        "operationId": "Venues_createAccessZone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccessZoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Venues"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ADMIN"
      }
    },
    "/api/v1/venues/{id}/gates": {
      "post": {
        "operationId": "Venues_createGate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Venues"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ADMIN"
      }
    },
    "/api/v1/gates/{id}": {
      "put": {
        "operationId": "Venues_updateGate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Venues"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ADMIN"
      }
    },
    "/api/v1/sections/{id}/seats/blocked": {
      "post": {
        "operationId": "Venues_setSeatsBlocked",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSeatsBlockedDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "401": {
            "description": "Missing or expired access token"
          },
          "403": {
            "description": "Signed in, but this role (or account) may not do this"
          }
        },
        "tags": [
          "Venues"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "description": "**Roles:** ADMIN"
      }
    },
    "/api/v1/events/{id}/seat-map": {
      "get": {
        "operationId": "Venues_seatMap",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Venues"
        ],
        "security": [
          {},
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Event Ticketing Platform API",
    "description": "REST API shared by the web app and the organizer/staff mobile apps. Amounts are integer minor units (butut, GMD). Sign in with POST /api/v1/auth/login, then send the access token as `Authorization: Bearer <token>`. See docs/api.md.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OfferTransferDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Who gets the ticket. They accept by signing in (or signing up) with this address.",
            "maxLength": 254
          }
        },
        "required": [
          "email"
        ]
      },
      "TransferTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The token from the emailed link (the part after #token=)",
            "minLength": 16,
            "maxLength": 200
          }
        },
        "required": [
          "token"
        ]
      },
...
      "CreateVenueDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "address",
          "city"
        ]
      },
      "SectionRowDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "pattern": "/^[A-Za-z0-9]{1,4}$/"
          },
          "seats": {
            "type": "number",
            "minimum": 1,
            "maximum": 200
          }
        },
        "required": [
          "label",
          "seats"
        ]
      },
      "CreateSectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "isVip": {
            "type": "boolean"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SectionRowDto"
            }
          }
        },
        "required": [
          "name",
          "rows"
        ]
      },
      "CreateAccessZoneDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "level": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "name",
          "level"
        ]
      },
      "CreateGateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "accessZoneId": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateGateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "accessZoneId": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SetSeatsBlockedDto": {
        "type": "object",
        "properties": {
          "seatIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isBlocked": {
            "type": "boolean"
          }
        },
        "required": [
          "seatIds",
          "isBlocked"
        ]
      },
...
    }
  }
}